// src/generated/t_tech/invest/grpc/instruments.ts
export interface FindInstrumentRequest {
  /** Строка поиска. */
  query: string;
  /** Фильтр по типу инструмента. */
  instrumentKind?:
    | InstrumentType
    | undefined;
  /** Фильтр для отображения только торговых инструментов. */
  apiTradeAvailableFlag?: boolean | undefined;
}

/** Результат поиска инструментов. */
export interface FindInstrumentResponse {
  /** Массив инструментов, удовлетворяющих условиям поиска. */
  instruments: InstrumentShort[];
}

/** Краткая информация об инструменте. */
export interface InstrumentShort {
  /** ISIN инструмента. */
  isin: string;
  /** FIGI инструмента. */
  figi: string;
  /** Ticker инструмента. */
  ticker: string;
  /** ClassCode инструмента. */
  classCode: string;
  /** Тип инструмента. */
  instrumentType: string;
  /** Название инструмента. */
  name: string;
  /** Уникальный идентификатор инструмента. */
  uid: string;
  /** Уникальный идентификатор позиции инструмента. */
  positionUid: string;
  /** Тип инструмента. */
  instrumentKind: InstrumentType;
  /** Возможность торговать инструментом через API. */
  apiTradeAvailableFlag: boolean;
  /** Признак доступности для ИИС. */
  forIisFlag: boolean;
  /** Дата первой минутной свечи. */
  first1minCandleDate:
    | Date
    | undefined;
  /** Дата первой дневной свечи. */
  first1dayCandleDate:
    | Date
    | undefined;
  /** Флаг, отображающий доступность торговли инструментом только для квалифицированных инвесторов. */
  forQualInvestorFlag: boolean;
  /** Флаг, отображающий доступность торговли инструментом по выходным. */
  weekendFlag: boolean;
  /** Флаг заблокированного ТКС. */
  blockedTcaFlag: boolean;
  /** Количество бумаг в лоте. */
  lot: number;
}
// ...
export const FindInstrumentRequest: MessageFns<FindInstrumentRequest> = {
  encode(message: FindInstrumentRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.query !== "") {
      writer.uint32(10).string(message.query);
    }
    if (message.instrumentKind !== undefined) {
      writer.uint32(16).int32(message.instrumentKind);
    }
    if (message.apiTradeAvailableFlag !== undefined) {
      writer.uint32(24).bool(message.apiTradeAvailableFlag);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): FindInstrumentRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseFindInstrumentRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.query = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.instrumentKind = reader.int32() as any;
          continue;
        }
        case 3: {
          if (tag !== 24) {
            break;
          }

          message.apiTradeAvailableFlag = reader.bool();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): FindInstrumentRequest {
    return {
      query: isSet(object.query) ? globalThis.String(object.query) : "",
      instrumentKind: isSet(object.instrumentKind)
        ? instrumentTypeFromJSON(object.instrumentKind)
        : isSet(object.instrument_kind)
        ? instrumentTypeFromJSON(object.instrument_kind)
        : undefined,
      apiTradeAvailableFlag: isSet(object.apiTradeAvailableFlag)
        ? globalThis.Boolean(object.apiTradeAvailableFlag)
        : isSet(object.api_trade_available_flag)
        ? globalThis.Boolean(object.api_trade_available_flag)
        : undefined,
    };
  },

  toJSON(message: FindInstrumentRequest): unknown {
    const obj: any = {};
    if (message.query !== "") {
      obj.query = message.query;
    }
    if (message.instrumentKind !== undefined) {
      obj.instrumentKind = instrumentTypeToJSON(message.instrumentKind);
    }
    if (message.apiTradeAvailableFlag !== undefined) {
      obj.apiTradeAvailableFlag = message.apiTradeAvailableFlag;
    }
    return obj;
  },

  create(base?: DeepPartial<FindInstrumentRequest>): FindInstrumentRequest {
    return FindInstrumentRequest.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<FindInstrumentRequest>): FindInstrumentRequest {
    const message = createBaseFindInstrumentRequest();
    message.query = object.query ?? "";
    message.instrumentKind = object.instrumentKind ?? undefined;
    message.apiTradeAvailableFlag = object.apiTradeAvailableFlag ?? undefined;
    return message;
  },
};

function createBaseFindInstrumentResponse(): FindInstrumentResponse {
  return { instruments: [] };
}

export const FindInstrumentResponse: MessageFns<FindInstrumentResponse> = {
  encode(message: FindInstrumentResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    for (const v of message.instruments) {
      InstrumentShort.encode(v!, writer.uint32(10).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): FindInstrumentResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseFindInstrumentResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.instruments.push(InstrumentShort.decode(reader, reader.uint32()));
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): FindInstrumentResponse {
    return {
      instruments: globalThis.Array.isArray(object?.instruments)
        ? object.instruments.map((e: any) => InstrumentShort.fromJSON(e))
        : [],
    };
  },

  toJSON(message: FindInstrumentResponse): unknown {
    const obj: any = {};
    if (message.instruments?.length) {
      obj.instruments = message.instruments.map((e) => InstrumentShort.toJSON(e));
    }
    return obj;
  },

  create(base?: DeepPartial<FindInstrumentResponse>): FindInstrumentResponse {
    return FindInstrumentResponse.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<FindInstrumentResponse>): FindInstrumentResponse {
    const message = createBaseFindInstrumentResponse();
    message.instruments = object.instruments?.map((e) => InstrumentShort.fromPartial(e)) || [];
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;
// ...
export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create(base?: DeepPartial<T>): T;
  fromPartial(object: DeepPartial<T>): T;
}